example avec des float

+++ main.cpp
@@ -13,43 +13,43 @@ using namespace coreutils;
 void testOp()
 {
 	/* Test Value */
-	ValueModel<int> *vm1 = new ValueModel<int>(3);
-	ValueModel<int> *vm2 = new ValueModel<int>(39);
+	ValueModel<float> *vm1 = new ValueModel<float>(3);
+	ValueModel<float> *vm2 = new ValueModel<float>(39);
 	cout << vm1->evaluate() << endl;
 	cout << vm2->evaluate() << endl;
 
 	/* Test Operator */
-	AggMax<int>* aggmax = new AggMax<int>();
-	AggPlus<int>* aggplus = new AggPlus<int>();
-	AndMin<int>* andmin = new AndMin<int>();
-	AndMult<int>* andmult = new AndMult<int>();
-	//CogDefuzz<int>* cogdefuzz = new CogDefuzz<int>();
-	NotMinus1<int>* notminus1 = new NotMinus1<int>();
-	OrMax<int>* ormax = new OrMax<int>();
-	OrPlus<int>* orplus = new OrPlus<int>();
-	//SugenoDefuzz<int>* sugenodefuzz = new SugenoDefuzz<int>();
-	ThenMin<int>* thenmin = new ThenMin<int>();
-	ThenMult<int>* thenmult = new ThenMult<int>();
+	AggMax<float>* aggmax = new AggMax<float>();
+	AggPlus<float>* aggplus = new AggPlus<float>();
+	AndMin<float>* andmin = new AndMin<float>();
+	AndMult<float>* andmult = new AndMult<float>();
+	//CogDefuzz<float>* cogdefuzz = new CogDefuzz<float>();
+	NotMinus1<float>* notminus1 = new NotMinus1<float>();
+	OrMax<float>* ormax = new OrMax<float>();
+	OrPlus<float>* orplus = new OrPlus<float>();
+	//SugenoDefuzz<float>* sugenodefuzz = new SugenoDefuzz<float>();
+	ThenMin<float>* thenmin = new ThenMin<float>();
+	ThenMult<float>* thenmult = new ThenMult<float>();
 
-	BinaryExpressionModel<int> a(vm1, vm2, aggmax);
+	BinaryExpressionModel<float> a(vm1, vm2, aggmax);
 	cout << a.evaluate() << endl;
-	BinaryExpressionModel<int> b(vm1, vm2, aggplus);
+	BinaryExpressionModel<float> b(vm1, vm2, aggplus);
 	cout << b.evaluate() << endl;
-	BinaryExpressionModel<int> c(vm1, vm2, andmin);
+	BinaryExpressionModel<float> c(vm1, vm2, andmin);
 	cout << c.evaluate() << endl;
-	BinaryExpressionModel<int> d(vm1, vm2, andmult);
+	BinaryExpressionModel<float> d(vm1, vm2, andmult);
 	cout << d.evaluate() << endl;
-	UnaryExpressionModel<int> e(vm1, notminus1);
+	UnaryExpressionModel<float> e(vm1, notminus1);
 	cout << e.evaluate() << endl;
-	BinaryExpressionModel<int> j(vm1, vm2, ormax);
+	BinaryExpressionModel<float> j(vm1, vm2, ormax);
 	cout << j.evaluate() << endl;
-	BinaryExpressionModel<int> g(vm1, vm2, orplus);
+	BinaryExpressionModel<float> g(vm1, vm2, orplus);
 	cout << g.evaluate() << endl;
-	BinaryExpressionModel<int> h(vm1, vm2, thenmin);
+	BinaryExpressionModel<float> h(vm1, vm2, thenmin);
 	cout << h.evaluate() << endl;
-	BinaryExpressionModel<int> i(vm1, vm2, thenmult);
+	BinaryExpressionModel<float> i(vm1, vm2, thenmult);
 	cout << i.evaluate() << endl;
-	//BinaryExpressionModel<int> j(vm1, vm2, orplus);
+	//BinaryExpressionModel<float> j(vm1, vm2, orplus);
 	//cout << j.evaluate() << endl;
 
 	delete vm1;
@@ -60,42 +60,42 @@ void testOp()
 void testShape()
 {
 	//utils
-	Interval<int> interval = Interval<int>(0, 30, 1);
+	Interval<float> interval = Interval<float>(0, 30, 1);
 
 	//Is
-	IsTriangle<int> tri = IsTriangle<int>(5, 10, 15);
+	IsTriangle<float> tri = IsTriangle<float>(5, 10, 15);
 
 }
 
 void testExample()
 {
 	//utils
-	Interval<int> interval = Interval<int>(0, 30, 1);
+	Interval<float> interval = Interval<float>(0, 30, 1);
 	//operators
-	NotMinus1<int> opNot;
-	AndMin<int> opAnd;
-	OrMax<int> opOr;
-	ThenMin<int> opThen;
-	AggMax<int> opAgg;
-	CogDefuzz<int> opDefuzz = CogDefuzz<int>(&interval);
+	NotMinus1<float> opNot;
+	AndMin<float> opAnd;
+	OrMax<float> opOr;
+	ThenMin<float> opThen;
+	AggMax<float> opAgg;
+	CogDefuzz<float> opDefuzz = CogDefuzz<float>(&interval);
 
 	//fuzzy expession factory
-	FuzzyFactory<int> f	(&opNot,&opAnd,&opOr,&opThen,&opAgg,&opDefuzz);
+	FuzzyFactory<float> f	(&opNot,&opAnd,&opOr,&opThen,&opAgg,&opDefuzz);
 
 	//membership function
-	IsTriangle<int> poor(-5,0,5);
-	IsTriangle<int> good(0,5,10);
-	IsTriangle<int> excellent(5,10,15);
-	IsTriangle<int> cheap(0,5,10);
-	IsTriangle<int> average(10,15,20);
-	IsTriangle<int> generous(20,25,30);
+	IsTriangle<float> poor(-5,0,5);
+	IsTriangle<float> good(0,5,10);
+	IsTriangle<float> excellent(5,10,15);
+	IsTriangle<float> cheap(0,5,10);
+	IsTriangle<float> average(10,15,20);
+	IsTriangle<float> generous(20,25,30);
 
 	//values
-	ValueModel<int> service(0);
-	ValueModel<int> food(5);
-	ValueModel<int> tips(0);
+	ValueModel<float> service(0);
+	ValueModel<float> food(5);
+	ValueModel<float> tips(0);
 
-	Expression<int> *r =
+	Expression<float> *r =
 			f.newAgg(
 					f.newAgg(
 							f.newThen(
@@ -114,7 +114,7 @@ void testExample()
 			);
 
 	//defuzzification
-	Expression<int> *system = f.newDefuzz(r,&tips);
+	Expression<float> *system = f.newDefuzz(r,&tips);
 
 	//apply input
 	float s;
